requirements.txt
streamlit
sentence-transformers
faiss-cpu
nbformat
pandas
openai
matplotlib
tqdm
